% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_funcs.R
\name{area_hex}
\alias{area_hex}
\title{Calculates the area of a hexagon given either the longer radius \eqn{R} or the apothem \eqn{r}}
\usage{
area_hex(R = waiver(), r = waiver())
}
\arguments{
\item{R}{The longer, or maximal, radius of a hexagon, from center to vertex Also the length of one of the sides.}

\item{r}{The apothem, or shorter radius of a hexagon, from center to side}
}
\value{
A numeric vector of areas, in the same units as \code{r} and/or \code{R}
}
\description{
from \url{https://stackoverflow.com/a/59022366}
}
\examples{
\dontrun{
 r <- 1.74 / 2
 area_hex(r = r)
 R <- 2 / 2
 area_hex(R = R)
 area_hex(R = R, r = r)
}
}
