% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ordered_script.R
\name{create_ordered_script}
\alias{create_ordered_script}
\title{Create a new script and numbers it with the first integer not used in
the directory. After creation the files is open for edition.}
\usage{
create_ordered_script(name, path = "R", edit_file = T)
}
\arguments{
\item{name}{(chr) The file name (without number)}

\item{path}{(chr) The target directory (default = "R")}

\item{edit_file}{(lgl) If TRUE the file will be open (default = T)}
}
\description{
Create a new script and numbers it with the first integer not used in
the directory. After creation the files is open for edition.
}
