% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_funcs.R
\name{decimal_places}
\alias{decimal_places}
\title{Counts the number of digits to the right of the decimal}
\usage{
decimal_places(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
An integer vector containing the number of digits to the right of the decimal for each
  element of \code{x}
}
\description{
Counts the number of digits to the right of the decimal place, ignoring any trailing zeroes.
}
\examples{
\dontrun{
 decimal_places(c(0.11, 23.4, 185.8987, 0.1100))
}
}
