% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_scripts.R
\name{delete_scripts}
\alias{delete_scripts}
\title{This function deletes any script that matches one of the patterns passed as
parameter.}
\usage{
delete_scripts(..., path = "R", confirm = T)
}
\arguments{
\item{...}{(chr) patterns used to select files for deletion}

\item{path}{(chr) The target directory (default = "R")}

\item{confirm}{(lgl) If FALSE no confirmation is asked (default = T)}
}
\description{
This function deletes any script that matches one of the patterns passed as
parameter.
}
