% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sci_notation.R
\name{exponent}
\alias{exponent}
\title{Exponent of a number in scientific notation}
\usage{
exponent(x)
}
\arguments{
\item{x}{(required) numeric. A number.}
}
\value{
the exponent of the scientific notation representation of the number \code{x}
}
\description{
Exponent of a number in scientific notation
}

