% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_funcs.R
\name{\%==\%}
\alias{\%==\%}
\title{Test if Two Numeric Vectors are (Nearly) Equal Row-by-Row}
\usage{
x \%==\% y
}
\arguments{
\item{x}{A numeric or integer vector. May be of length 1.}

\item{y}{A numeric or integer vector to compare to x. May be of length 1.}
}
\value{
A logical vector of TRUE and FALSE values indicating which rows
  are (nearly) equal.
}
\description{
Implements the base R function \link{all.equal} on a row-by-row basis.
  x and y can be of different lengths, including \code{length() == 1}.
  Handy for making comparisons between data frame columns
  within \link[dplyr]{filter} and similar functions.
}
\examples{
\dontrun{
 x <- 1L
 y <- 1.001
 x == y
 # FALSE
 x \%==\% y
 # FALSE
 y <- 1.00000001
 x == y
 # FALSE
 x \%==\% y
 # TRUE
 x <- c(1, 2)
 x == y
 # FALSE FALSE
 x \%==\% y
 # TRUE FALSE
}
}
