% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_script_at.R
\name{insert_script_at}
\alias{insert_script_at}
\title{Create a new script and numbers it with the first integer not used in
the directory. After creation the files is open for edition.}
\usage{
insert_script_at(path = "R", pos, name)
}
\arguments{
\item{path}{(chr) The target directory (default = "R")}

\item{pos}{(int) The position at which the new file will be inserted}

\item{name}{(chr) The name of the new files (without number)}
}
\description{
Create a new script and numbers it with the first integer not used in
the directory. After creation the files is open for edition.
}
