% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_funcs.R
\name{is.int}
\alias{is.int}
\title{Determins if a number is an IEEE-754 integer}
\source{
\url{http://www.win-vector.com/blog/2015/06/r-in-a-64-bit-world/}
}
\usage{
is.int(x, ...)
}
\arguments{
\item{x}{the vector to check}

\item{...}{provides backward compatibility with previous versions.}
}
\value{
Returns TRUE for each element of \code{x} that is an IEEE-754-compliant integer, and FALSE for all other elements.
}
\description{
Determines if a vector fits the IEEE-754 definition of an integer that can be exactly represented with a double-precision number.
}
\examples{
 x <- c(1, 412, 4.4, 0.003)
 is.int(x)
}
