% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci_notation.R
\name{mantissa}
\alias{mantissa}
\title{Mantissa of a number in scientific notation}
\usage{
mantissa(x)
}
\arguments{
\item{x}{(required) numeric. A number.}
}
\value{
the mantissa of the scientific notation representation of the number \code{x}
}
\description{
Returns the mantissa of a number as it is written in scientific
  notation (powers of 10).
}
\examples{
 x <- c(1,400,500000,.0003)
 mantissa(x)
}
