% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_funcs.R
\name{mean_geom}
\alias{mean_geom}
\title{Returns the geometric mean of a vector}
\usage{
mean_geom(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric or integer vector}

\item{na.rm}{a boolean indicating, if \code{TRUE}, removes \code{NA} values before calculating the mean}
}
\value{
The geometric mean of the vector \code{x}, or \code{NULL} with a warning
  if the geometric mean cannot be calculated
}
\description{
Calculates the geometric mean of a vector, returning a single value. The geometric mean,
  like the mean and the median, is an indication of the central tendancy of a set of numbers.
  The geometric mean is  the \emph{n}th root of the product of \emph{n} numbers. For instance,
  the geometric mean of 2 and 8 is \code{sqrt(2 * 8) = 4}.
  The geometric mean is useful when computing the central tendency of measures that have different
  ranges. For instance, when computing a single "figure of merit" from differentrating scales that
  have ranges 0 to 5 and 0 to 100.
}
