% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_funcs.R
\name{mode_stat}
\alias{mode_stat}
\title{Returns the mode of a vector.}
\usage{
mode_stat(x, na.rm = FALSE)
}
\arguments{
\item{x}{a vector to determine the mode of}

\item{na.rm}{a boolean indicating, if \code{TRUE}, removes \code{NA} values before calculating the mode}
}
\value{
a vector containing the modes of \code{x}
}
\description{
Determines the most common value in a vector. Handles multiple modes by returning a vector
 containing all modes. Works with any data type.
}
