% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbr_insert_at.R
\name{nbr_insert_at}
\alias{nbr_insert_at}
\title{Create a new script and numbers it with the integer used as parameter.
All files are renumbered accordingly the directory. After creation the file
is open for edition.}
\usage{
nbr_insert_at(path = "R", pos, name, edit_file = T)
}
\arguments{
\item{path}{(chr) The target directory (default = "R")}

\item{pos}{(int) The position at which the new file will be inserted}

\item{name}{(chr) The name of the new files (without number)}

\item{edit_file}{(lgl) If TRUE the file will be open (default = T)}
}
\description{
Create a new script and numbers it with the integer used as parameter.
All files are renumbered accordingly the directory. After creation the file
is open for edition.
}
