% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbr_source.R
\name{nbr_source}
\alias{nbr_source}
\title{Source script and all previous script that have been modified
since last time they were sourced.}
\usage{
nbr_source(script, dir = "R", confirm_duration = 15 * 1000)
}
\arguments{
\item{script}{(chr) The name of the script file}

\item{dir}{(chr) The target directory (default = "R")}

\item{confirm_duration}{(int) duration threshold to ask for confirmation (ms)}
}
\description{
Source script and all previous script that have been modified
since last time they were sourced.
}
