% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbr_source_all.R
\name{nbr_source_all}
\alias{nbr_source_all}
\title{Run all numbered scripts in the target directory.}
\usage{
nbr_source_all(path = "R")
}
\arguments{
\item{path}{(chr) The target directory (default = "R")}
}
\description{
Run all numbered scripts in the target directory.
}
