% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_funcs.R
\name{nCr}
\alias{nCr}
\title{Calculate the number of possible combinations of \eqn{n} objects, taken \eqn{r} at a time without repitition}
\usage{
nCr(n, r)
}
\arguments{
\item{n}{An integer number of elements to choose from for the combination. May be a vector if \code{r} has a single element (i.e. \code{length(r) == 1}).}

\item{r}{An integer subgroup size to combine. May be a vector if \code{n} has a single element (i.e. \code{length(n) == 1}).}
}
\value{
An integer number of possible permutations
}
\description{
Calculates the number of arrangements in which no element occurs more than once and order does
  not matter, without the requirement of using all the elements from a given set. For example, if we were
  arranging an apple, a pear, and an orange (\eqn{n = 3}) into sets of two (\eqn{k = 2}), we would find
  that there are three possible combinations: an apple and a pear, an apple and an orange, and a pear
  and an orange. We would not count combinations where the order was reversed (e.g. a pear and an apple)
  as different combinations.
}
\details{
Implements the equation \deqn{ \frac{n!}{r!(n - r)!} }{n! / (r! * (n - r)!)}
}
\examples{
  \dontrun{
    nCr(3, 2)
    nCr(3:5, 2)
  }
}
