% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.R
\name{nin}
\alias{nin}
\alias{\%nin\%}
\title{Not in operator}
\usage{
x \%nin\% y
}
\arguments{
\item{x}{a value to check}

\item{y}{a value or vector of values to check against x}
}
\value{
Boolean value, true if x is not in y; false otherwise
}
\description{
Checks if x is in y
}
