% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_funcs.R
\name{nPr}
\alias{nPr}
\title{Calculate the number of possible partial permutations of \eqn{n} objects, \eqn{r} at a time}
\usage{
nPr(n, r)
}
\arguments{
\item{n}{An integer number of elements to permute. May be a vector if \code{r} has a single element (i.e. \code{length(r) == 1}).}

\item{r}{An integer subgroup size. May be a vector if \code{n} has a single element (i.e. \code{length(n) == 1}).}
}
\value{
An integer number of possible permutations
}
\description{
Calculates the number of ordered arrangements in which no element occurs more than once,
  without the requirement of using all the elements from a given set. Also known as \emph{partial permutations}
   or as \emph{sequences without repetition}.
}
\details{
Implements the equation \deqn{\frac{n!}{(n - r)!}}{n! / (n - r)!}
}
\examples{
  \dontrun{
    nPr(3, 2)
    nPr(3:5, 2)
  }
}
