% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbr-package.R
\docType{package}
\name{numbr-package}
\alias{numbr}
\alias{numbr-package}
\title{numbr: Number Your Scripts}
\description{
Organize your R script directory with numbered files.
}
\author{
\strong{Maintainer}: Pierre Balaye \email{pierrebalaye@gmail.com} (\href{https://orcid.org/0000-0002-7045-5065}{ORCID})

}
\keyword{internal}
