% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_funcs.R
\name{pred_r_squared}
\alias{pred_r_squared}
\title{Predicted R squared}
\usage{
pred_r_squared(linear.model)
}
\arguments{
\item{linear.model}{An \code{lm} or \code{glm} object}
}
\value{
The predicted r squared value
}
\description{
Calculates the predicted r-squared statistic.
		Used primarily to compare alternative models.
}
