% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_funcs.R
\name{PRESS}
\alias{PRESS}
\title{PRESS statistic}
\usage{
PRESS(linear.model)
}
\arguments{
\item{linear.model}{An \code{lm} or \code{glm} object}
}
\value{
The PRESS statistic
}
\description{
Calculates the predicted residual sum of squares, or PRESS, statistic.
		Used primarily to compare alternative models.
}
