% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_numbered_scripts.R
\name{run_numbered_scripts}
\alias{run_numbered_scripts}
\title{Run all numbered scripts in the target directory.}
\usage{
run_numbered_scripts(path = "R")
}
\arguments{
\item{path}{(chr) The target directory (default = "R")}
}
\description{
Run all numbered scripts in the target directory.
}
