% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_scripts.R
\name{delete_scripts}
\alias{delete_scripts}
\title{Create a new script and numbers it with the first integer not used in
the directory. After creation the files is open for edition.}
\usage{
delete_scripts(..., path = "R", confirm = T)
}
\arguments{
\item{...}{(chr) patterns used to select files for deletion}

\item{path}{(chr) The target directory (default = "R")}

\item{confirm}{(lgl) If FALSE no confirmation is asked (default = T)}
}
\description{
Create a new script and numbers it with the first integer not used in
the directory. After creation the files is open for edition.
}
