% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sci_notation.R
\name{mantissa}
\alias{mantissa}
\title{Mantissa of a number in scientific notation}
\usage{
mantissa(x)
}
\arguments{
\item{x}{(required) numeric. A number.}
}
\value{
the mantissa of the scientific notation representation of the number \code{x}
}
\description{
Mantissa of a number in scientific notation
}

