% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_funcs.R
\name{mode_stat}
\alias{mode_stat}
\title{Returns the mode of a vector.}
\usage{
mode_stat(x, na.rm = FALSE, method = "tabulate")
}
\arguments{
\item{x}{a vector}

\item{na.rm}{a boolean indicating, if \code{TRUE}, removes \code{NA} values before calculating the mode}

\item{method}{specifies the method used for determining the mode}
}
\value{
a vector containing the modes of \code{x}
}
\description{
Determines the most common value in a vector. Handles multiple modes by returning a vector
 containing all modes. Works with any data type.
}
\details{
\code{method} must be one of either "tabulate" or "density." "Tabulate"
}
