% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbr_source_all.R
\name{nbr_source_all}
\alias{nbr_source_all}
\title{Run all numbered scripts in the target directory.}
\usage{
nbr_source_all(dir = "R", confirm_duration = 10 * 1000)
}
\arguments{
\item{dir}{(chr) The target directory (default = "R")}

\item{confirm_duration}{(int) duration threshold to ask for confirmation (ms)}
}
\description{
Run all numbered scripts in the target directory.
}
