% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_funcs.R
\name{num_order_to_word}
\alias{num_order_to_word}
\title{Convert a number to large-number word representation}
\usage{
num_order_to_word(x, lookup = NULL)
}
\arguments{
\item{x}{A number to convert. length(x) must equal 1.}

\item{lookup}{A data frame specifying numeric exponents as `expon` and corresponding names as `word`. e.g. lookup = data.frame(expon = c(3, 0, -3), word = c("thousands", "", "thousandths"))}
}
\value{
A list containing the rounded number and its string representation
}
\description{
Converts a number to a character string approximation using the "short scale"
  version of large number names. e.g. 312e6 returns as '300 million.'
  Simultaneously returns a numeric representation of the approximation.
}
\examples{
# Simplest example
num_order_to_word(1340)
# Using a variable
x <- exp(1)^20
num_order_to_word(x)
x <- exp(1)^-10
num_order_to_word(x)
# Fails with an error message
\dontrun{
x <- 1:4
num_order_to_word(x)
}
}
